Assertions, warning suppression, minor cleanups, and (void)::CloseHandle pocesshandle after CreatePRocess to prevent leak

+++ Tests/37/Test.cpp
@@ -51,7 +51,7 @@ namespace   {
         wchar_t outBuf[10 * 1024];
 
         String canonical;
-        ThrowIfErrorHRESULT (::CoInternetParseUrl (CComBSTR (w.c_str ()), PARSE_CANONICALIZE, 0, outBuf, NEltsOf (outBuf), &ingored, 0));
+        ThrowIfErrorHRESULT (::CoInternetParseUrl (CComBSTR (w.c_str ()), PARSE_CANONICALIZE, 0, outBuf, static_cast<DWORD> (NEltsOf (outBuf)), &ingored, 0));
         canonical = outBuf;
 
         {
@@ -61,7 +61,7 @@ namespace   {
             }
         }
 
-        if (SUCCEEDED (::CoInternetParseUrl (CComBSTR (canonical.c_str ()), PARSE_DOMAIN, 0, outBuf, NEltsOf (outBuf), &ingored, 0))) {
+        if (SUCCEEDED (::CoInternetParseUrl (CComBSTR (canonical.c_str ()), PARSE_DOMAIN, 0, outBuf, static_cast<DWORD> (NEltsOf (outBuf)), &ingored, 0))) {
             *host = outBuf;
         }
